Add SEO: per-page metadata, OG tags, sitemap, robots.txt, JSON-LD
- Split all 'use client' pages into server wrapper + client component
  so Next.js can export metadata per route
- Per-page title, description, OG, and Twitter card tags
- LocalBusiness JSON-LD schema in layout
- /sitemap.xml and /robots.txt via Next.js file-based routes
- metadataBase + title template set on layout
