+++ app/sitemap.ts
@@ -1,0 +1,12 @@
+import type { MetadataRoute } from 'next';
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  const base = 'https://contraptionsoft.com';
+  return [
+    { url: base,               lastModified: new Date(), changeFrequency: 'monthly', priority: 1 },
+    { url: `${base}/products`, lastModified: new Date(), changeFrequency: 'monthly', priority: 0.8 },
+    { url: `${base}/work`,     lastModified: new Date(), changeFrequency: 'monthly', priority: 0.8 },
+    { url: `${base}/about`,    lastModified: new Date(), changeFrequency: 'monthly', priority: 0.7 },
+    { url: `${base}/contact`,  lastModified: new Date(), changeFrequency: 'yearly',  priority: 0.6 },
+  ];
+}
